requirements.txt
streamlit
sqlalchemy
python-dotenv
mistralai
Pillow
pymupdf